% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwutilityfunctions.R
\name{analyze_cosinor_design}
\alias{analyze_cosinor_design}
\title{Analyze Cosinor Design}
\usage{
analyze_cosinor_design(timepoints, per = 24, accuracy = 1e-04)
}
\arguments{
\item{timepoints}{A numeric vector of time points to analyze.}

\item{per}{The cosinor period}

\item{accuracy}{Absolute accuracy for determining centeredness and phase
invariance}
}
\value{
An object of class \code{cosinor_design} containing the following 
elements:
\describe{
  \item{n_points}{The total number of time points.}
  \item{n_uniq}{The number of unique time points (sorted).}
  \item{reps}{A vector indicating the number of replicates per unique time 
  point (sorted).}
  \item{reps_uniq}{A vector of unique replicate counts across all time 
  points.}
  \item{n_unique_modper}{The number of unique time points within a per-hour 
  cycle.}
  \item{reps_modper}{A vector of replicate counts per unique time point 
  modulo per.}
  \item{reps_modper_uniq}{A vector of unique replicate counts across all time
  points modulo per.}
  \item{sampling_int_uniq}{A vector of unique sampling intervals between 
  consecutive time points (sorted).}
  \item{tps_sorted}{A vector of the input time points sorted in ascending 
  order.}
  \item{tps_modper_sorted}{A vector of the time points modulo per, sorted.}
  \item{input_sort_idx}{The index mapping the original input time points to 
  the sorted time points.}
  \item{input_to_modper_idx}{The index mapping the original input time points
  to the sorted modulo-per time points.}
  \item{equal_reps}{Logical, indicating whether the number of replicates
  per unique time point modulo per is identical.}
  \item{evenly_spaced}{Logical, indicating whether time points are evenly
  spaced within a cycle, also considering wrap-around from end to beginning.}
  \item{is_centered}{Logical, indicating whether the design is centered.
  Centered designs potentially reach the highest power.}
  \item{is_phaseinv}{Logical, indicating whether the design is
  phase-invariant. Phase-invariant designs lead to power independent of
  phase.}
}
}
\description{
This function analyzes a vector of time points and computes various metrics
related to periodic sampling. It handles both absolute and modulo-per time
representations.
}
\details{
The function performs several analyses on the provided time points:
\itemize{
  \item Computes the number of replicates for each unique time point.
  \item Computes the number of unique time points and replicates modulo per.
  \item Checks for balance in replicates, sampling intervals, and 
  distribution within a cycle.
}
}
\examples{
# Simulate a time point vector
set.seed(123)
timepoints <- rep(seq(0, 44, by = 4), each = 3) |> sample()

# Analyze time points
results <- analyze_cosinor_design(timepoints)

# View results
print(results)

}
