% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwutilityfunctions.R
\name{analyze_zt}
\alias{analyze_zt}
\title{Analyze Zeitgeber Time Points}
\usage{
analyze_zt(timepoints)
}
\arguments{
\item{timepoints}{A numeric vector of time points to analyze.}
}
\value{
An object of class \code{zt_analysis} containing the following 
elements:
\describe{
  \item{n_points}{The total number of time points.}
  \item{n_uniq}{The number of unique time points (sorted).}
  \item{reps}{A vector indicating the number of replicates per unique time 
  point (sorted).}
  \item{reps_uniq}{A vector of unique replicate counts across all time 
  points.}
  \item{n_unique_mod24}{The number of unique time points within a 24-hour 
  cycle.}
  \item{reps_mod24}{A vector of replicate counts per unique time point 
  modulo 24.}
  \item{reps_mod24_uniq}{A vector of unique replicate counts across all time
  points modulo 24.}
  \item{sampling_int_uniq}{A vector of unique sampling intervals between 
  consecutive time points (sorted).}
  \item{tps_sorted}{A vector of the input time points sorted in ascending 
  order.}
  \item{tps_mod24_sorted}{A vector of the time points modulo 24, sorted.}
  \item{input_sort_idx}{The index mapping the original input time points to 
  the sorted time points.}
  \item{input_to_mod24_idx}{The index mapping the original input time points
  to the sorted modulo-24 time points.}
  \item{balanced_reps}{Logical, indicating whether the number of replicates
  per unique time point modulo 24 is identical.}
  \item{balanced}{Logical, indicating whether the design is balanced overall
  (based on replicates, intervals, and distribution).}
}
}
\description{
This function analyzes a vector of time points and computes various metrics
related to periodic sampling. It handles both absolute and modulo-24 time
representations, making it suitable for circadian rhythm analyses.
}
\details{
The function performs several analyses on the provided time points:
\itemize{
  \item Computes the number of replicates for each unique time point.
  \item Computes the number of unique time points and replicates modulo 24.
  \item Checks for balance in replicates, sampling intervals, and 
  distribution within a 24-hour cycle.
}

  The function is especially useful for analyzing time points in circadian
  rhythm studies, where time points within a 24-hour cycle are often treated
  as equivalent.
}
\examples{
# Simulate a time point vector
set.seed(123)
timepoints <- rep(seq(0, 44, by = 4), each = 3) |> sample()

# Analyze time points
results <- analyze_zt(timepoints)

# View results
print(results)

}
