% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwutilityfunctions.R
\name{char2sorted_factor}
\alias{char2sorted_factor}
\title{Converts character vector to sorted factor}
\usage{
char2sorted_factor(charvec, ...)
}
\arguments{
\item{charvec}{A character vector}

\item{...}{Further arguments to gtools::mixedsort}
}
\value{
Sorted factor
}
\description{
Sorting is carried out using "human" criteria with gtools::mixedsort.
Useful with e.g., ggplot or tidyr::spread
}
\examples{
df <- dplyr::data_frame(cells=paste0("cell", 1:15), vals=1:15)
df \%>\% tidyr::spread(cells, vals)
df \%>\% dplyr::mutate(cells=char2sorted_factor(cells)) \%>\%
tidyr::spread(cells, vals)

}
