% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwutilityfunctions.R
\name{circular_spectral_colors}
\alias{circular_spectral_colors}
\title{Computes color components along a circular spectrum}
\usage{
circular_spectral_colors(ncolor = 24L, ncolor.orig = 11L)
}
\arguments{
\item{ncolor}{Number of desired colors}

\item{ncolor.orig}{Number of colors in the original spectrum, defaults to 11,
the maximal number of available colors in RColorBrewer}
}
\value{
Desired colors in hex format
}
\description{
Uses RColorBrewer as basis
}
\examples{
graphics::pie(rep(1, 24), col=circular_spectral_colors(24L))
}
