% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwutilityfunctions.R
\name{fisher_test_df}
\alias{fisher_test_df}
\title{Performs Fisher test for odds difference of condition 1 contingent on
condition 2.}
\usage{
fisher_test_df(df, cond1, cond2, alternative = "two.sided", ...)
}
\arguments{
\item{df}{tibble}

\item{cond1}{quosure evaluating to a logical for condition 1}

\item{cond2}{quosure evaluating to a logical for condition 2}

\item{alternative}{default="two.sided", see fisher.test().}

\item{...}{further arguments to \code{my.fisher.test.num()} such as
\code{verbose}}
}
\value{
p value and odds ratio (c1c2/!c1c2) / (c1!c2/!c1!c2)
}
\description{
NAs must be taken into account when formulating the conditions below,
if desired.
}
\details{
The matrix (returned if \code{verbose=TRUE}) is constructed with cond1
and !cond1 for the first and second row, respectively; likewise with cond2
and !cond2 for the first and second column, respectively
}
\examples{
df <- data_frame(a=1:10, b=4:13)
fisher_test_df(df, quo(a < 4), quo(b < 7))
fisher_test_df(df, quo(a > 2), quo(b < 8))
}
