% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwutilityfunctions.R
\name{forge_polar_amp_df}
\alias{forge_polar_amp_df}
\title{Constructs data_frame suitable for polar circadian phase arrows}
\usage{
forge_polar_amp_df(df, amp_mean, phase_mean, grp = rlang::quo(what))
}
\arguments{
\item{df}{input data_frame with:}

\item{amp_mean}{amplitudes (quosure or values overriding those in data frame)}

\item{phase_mean}{phases (quosure, see above)}

\item{grp}{grouping variable (quosure)}
}
\value{
data_frame suitable for ggplot
}
\description{
Constructs data_frame suitable for polar circadian phase arrows
}
\examples{
df <- dplyr::data_frame(what="test", amp=0.5, phase=12)
df \%>\% forge_polar_amp_df(rlang::quo(amp), rlang::quo(phase),
rlang::quo(what))
}
