% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwutilityfunctions.R
\name{forge_polar_phaseint_df}
\alias{forge_polar_phaseint_df}
\title{Constructs data_frame suitable for polar circadian phase intervals}
\usage{
forge_polar_phaseint_df(df, rad, phase_mean, phase_int, grp = rlang::quo(what))
}
\arguments{
\item{df}{input data_frame with:}

\item{rad}{radial distance(s) for interval segment (quosure)}

\item{phase_mean}{phase means (quosure)}

\item{phase_int}{e.g. phase SD, positive (quosure)}

\item{grp}{grouping variable (quosure)}
}
\value{
data_frame suitable for ggplot
}
\description{
Constructs data_frame suitable for polar circadian phase intervals
}
\examples{
df <- dplyr::data_frame(what="test", rad=1, phase_mean=23, phase_int=2)
df \%>\% forge_polar_phaseint_df(rlang::quo(rad),
rlang::quo(phase_mean), rlang::quo(phase_int))
}
