% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwutilityfunctions.R
\name{gene_names_to_entrez}
\alias{gene_names_to_entrez}
\title{Search Entrez IDs for gene names that are not necessarily MGI symbols}
\usage{
gene_names_to_entrez(gene_names)
}
\arguments{
\item{gene_names}{Character vector of gene names}
}
\value{
Data frame (tibble) with columns "alias" and "entrez_id" for unique
  matches
}
\description{
Search Entrez IDs for gene names that are not necessarily MGI symbols
}
\examples{
gene_names_to_entrez(c("Bmal1", "Ciart", "Chrono"))
}
