% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwutilityfunctions.R
\name{interleave_index}
\alias{interleave_index}
\title{Interleave columns of a matrix based on specified column indices}
\usage{
interleave_index(matrix_data, ..., along_columns = TRUE)
}
\arguments{
\item{matrix_data}{matrix to interleave}

\item{...}{column index vectors to extract and interleave}

\item{use_columns}{logical, whether to transpose the matrix before and after,
i.e., operate on columns instead of rows}
}
\value{
Interleaved matrix
}
\description{
Interleave columns of a matrix based on specified column indices
}
\examples{
a <- matrix(1:8, ncol = 4)
interleave_index(a, 1:2, 3:4)

}
