% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwutilityfunctions.R
\name{interleave_segments}
\alias{interleave_segments}
\title{Interleave segments of a matrix based on specified indices}
\usage{
interleave_segments(matrix_data, ..., use_columns = TRUE)
}
\arguments{
\item{matrix_data}{A matrix to interleave}

\item{...}{Index vectors to extract segments (columns or rows) to interleave}

\item{use_columns}{Logical, whether to operate on columns. If TRUE, the 
matrix is transposed before and after interleaving.}
}
\value{
Interleaved matrix
}
\description{
Interleave segments of a matrix based on specified indices
}
\examples{
a <- matrix(1:8, ncol = 4)
interleave_segments(a, 1:2, 3:4, use_columns = TRUE)
interleave_segments(a, 1:2, 3:4, use_columns = FALSE)
}
