% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwutilityfunctions.R
\name{inverse}
\alias{inverse}
\title{Create inverse function}
\usage{
inverse(fn, interval = NULL, lower = min(interval), upper = max(interval), ...)
}
\arguments{
\item{fn}{Monotonous function}

\item{interval}{Vector defining the domain}

\item{lower}{Numeric, inferred from interval or given explicitly here}

\item{upper}{Numeric, inferred from interval or given explicitly here}

\item{...}{Further parameters to uniroot()}
}
\value{
Function (vectorized) computing the inverse of fn
}
\description{
From
https://stackoverflow.com/questions/45616072/find-the-inverse-function-in-r
}
\examples{
x <- 1:10
y <- sqrt(x)
sqrt.inv <- inverse(sqrt, lower = 1, upper = 10)
sqrt.inv(y)
}
