% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwutilityfunctions.R
\name{make_index_vector}
\alias{make_index_vector}
\title{Make a vector of stretches of indices of contigs satisfying a predicate
vector}
\usage{
make_index_vector(vec, predicate = !is.na(vec))
}
\arguments{
\item{vec}{A vector}

\item{predicate}{A TRUE/FALSE vector of same length as vec, defaults to
!is.na(vec)}
}
\value{
vector of indices
}
\description{
Useful together with split(), see example
}
\examples{
tosplit <- c(TRUE, TRUE, TRUE, NA, NA, TRUE, TRUE, NA,
TRUE, TRUE, TRUE, TRUE)
(index <- make_index_vector(tosplit))
split(tosplit[!is.na(tosplit)], index)

set.seed(1234)
(tosplit2 <- rnorm(10))
make_index_vector(tosplit2, tosplit2 > 0)

}
