% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwutilityfunctions.R
\name{matrixslices}
\alias{matrixslices}
\title{Matrix columnwise slice centered on different positions for each column}
\usage{
matrixslices(mat, positions, posminus, posplus)
}
\arguments{
\item{mat}{A matrix}

\item{positions}{Center positions, vector. One element for each column in mat}

\item{posminus}{Scalar number of upstream positions in addition to center}

\item{posplus}{Scalar number of downstream positions in addition to center}
}
\value{
Matrix of column slices
}
\description{
Matrix columnwise slice centered on different positions for each column
}
\examples{
A <- matrix(1:10, ncol=2)
matrixslices(A, c(4, 2), 1, 1)

}
