% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwutilityfunctions.R
\name{my_intronsByGene2}
\alias{my_intronsByGene2}
\title{List of genes with introns}
\usage{
my_intronsByGene2(txdb)
}
\arguments{
\item{txdb}{TxDb database object}
}
\value{
GRangesList with flattened genes with introns
}
\description{
List of genes with introns
}
\examples{
library(TxDb.Mmusculus.UCSC.mm9.refGene)
my_intronsByGene(TxDb.Mmusculus.UCSC.mm9.refGene)

}
\references{
See post on (https://stat.ethz.ch/pipermail/
bioconductor/2013-February/050725.html)
}
