% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwutilityfunctions.R
\name{myfrac}
\alias{myfrac}
\title{Print the fraction of TRUEs in a vector of logicals}
\usage{
myfrac(vec)
}
\arguments{
\item{vec}{Vector of logicals}
}
\value{
Fraction of TRUE values in vec
}
\description{
Print the fraction of TRUEs in a vector of logicals
}
\examples{
set.seed(123)
myfrac(rnorm(10) > 0)
myfrac(TRUE)
myfrac(FALSE)
}
