% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwutilityfunctions.R
\name{mysignedcircpairdist.ct}
\alias{mysignedcircpairdist.ct}
\title{Signed differences in phase (0-24 hrs scale)}
\usage{
mysignedcircpairdist.ct(x, y)
}
\arguments{
\item{x}{Vector, phases between 0 and 24}

\item{y}{Vector, phases between 0 and 24}
}
\value{
Vector of signed phase differences
}
\description{
Computes phase difference between x and y, between -12 and 12.
A positive result means that x has earlier phase than y. See example.
}
\examples{
mysignedcircpairdist.ct(c(10, 22, 2), c(12, 2, 23))
}
