% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwutilityfunctions.R
\name{qmetade}
\alias{qmetade}
\title{Perform p value meta analysis using MetaDE}
\usage{
qmetade(..., metameth = "minP")
}
\arguments{
\item{...}{2 or more vectors of equal lengths of p values}

\item{metameth}{meta analysis method, defaluts to minP
(OR=union-intersection)}
}
\value{
vector of p values
}
\description{
Perform p value meta analysis using MetaDE
}
\examples{
qmetade(runif(10), runif(10))
}
