% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwutilityfunctions.R
\name{reduce_gene_table_ambiguity}
\alias{reduce_gene_table_ambiguity}
\title{Remove some ambiguous entries in a gene translation data frame}
\usage{
reduce_gene_table_ambiguity(gene_table, id1, id2)
}
\arguments{
\item{gene_table}{Data frame}

\item{id1}{Name of column to split (unquoted)}

\item{id2}{Index column for the split}
}
\value{
Reduced data frame
}
\description{
Assuming a data frame with at least 2 columns, we may split 1 column on the
other. We may then remove entries in multiple-entry elements that occur in
1-entry elements - these are likely mis-assigned annotations.
}
\examples{
test <- data.frame(ida = c("a", "b", "a"), idb = c("x", "x", "y"))
reduce_gene_table_ambiguity(test, ida, idb)

}
