% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwutilityfunctions.R
\name{roundn}
\alias{roundn}
\title{Round to nearest integer multiple of n}
\usage{
roundn(x, n = 1)
}
\arguments{
\item{x}{A vector}

\item{n}{A number}
}
\value{
Rounded vector
}
\description{
Round to nearest integer multiple of n
}
\examples{
roundn(runif(10)*10, 2)
}
