% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwutilityfunctions.R
\name{running_average}
\alias{running_average}
\title{Running average}
\usage{
running_average(vec, winlength, keep.na = TRUE)
}
\arguments{
\item{vec}{Vector to perform running average on}

\item{winlength}{Averaging window length}

\item{keep.na}{Should NAs at beginning and end be kept?
(They start at the end)}
}
\value{
Running average
}
\description{
Running average
}
\examples{
vec <- 1:4
running_average(vec, 2)
running_average(vec, 3)
running_average(vec, 2, keep.na=FALSE)
}
