% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwutilityfunctions.R
\name{sel_padj}
\alias{sel_padj}
\title{Compute adjusted p values with pre-filtering}
\usage{
sel_padj(pvals, ind, method = "BH")
}
\arguments{
\item{pvals}{vector of input p values}

\item{ind}{logical vector corresponding to pre-filtering condition. NAs are
allowed and reinterprated as FALSE.}

\item{method}{method selection for p.adjust}
}
\value{
vector of adjusted p values with NAs for non-satisfied pre-filtering
}
\description{
Compute adjusted p values with pre-filtering
}
\examples{
pvals <- runif(10)
effectsize <- runif(10)
sel_padj(pvals, effectsize > 0.5)
}
