% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwutilityfunctions.R
\name{sel_qval}
\alias{sel_qval}
\title{Compute q values with pre-filtering}
\usage{
sel_qval(pvals, ind = NULL, method = "bootstrap", ...)
}
\arguments{
\item{pvals}{vector of input p values}

\item{ind}{logical vector corresponding to pre-filtering condition. NAs are
allowed and reinterprated as FALSE. If NULL, all are interpreted as TRUE.}

\item{method}{method selection for qvalue::pi0est}

\item{...}{passed on to qvalue::qvalue (and then to pi0est)}
}
\value{
vector of q values with NAs for non-satisfied pre-filtering
}
\description{
Compute q values with pre-filtering
}
\examples{
pvals <- runif(10)
effectsize <- runif(10)
sel_qval(pvals, effectsize > 0.5)
}
