% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwutilityfunctions.R
\name{shifter}
\alias{shifter}
\title{Shifts vector by cyclic permutation}
\usage{
shifter(vec, n = 1L)
}
\arguments{
\item{vec}{vector to cycle}

\item{n}{amount of cyclic permutation}
}
\value{
shifted vector
}
\description{
Shifts vector by cyclic permutation
}
\examples{
shifter(1:5, 2)
shifter(1:5, -2)
}
\references{
https://stackoverflow.com/questions/30542128/
circular-shifting-arrays-in-r-by-distance-n
}
