% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwutilityfunctions.R
\name{split024intervals}
\alias{split024intervals}
\title{Split data frame containing ZTs over the 24/0 border}
\usage{
split024intervals(
  df,
  grp = rlang::quo(what),
  x = "x",
  y = "y",
  xend = "xend",
  yend = "yend"
)
}
\arguments{
\item{df}{Data frame to split}

\item{grp}{Grouping variable (quosure), split will be made for each group}

\item{x}{Starting points, secondary variable in polar coordinates, e.g. length
or amplitude}

\item{y}{Primary polar variable, e.g. time of day, ZT}

\item{xend}{Corresponding x end points}

\item{yend}{y end points}
}
\value{
Split data frame
}
\description{
Split data frame containing ZTs over the 24/0 border
}
\examples{
# From plot in our paper 10.1073/pnas.1613103114
PFK <- 10.626
PFKup <- PFK + 2.9822/2
PFKdown <- PFK - 2.9822/2
ciopt <- dplyr::data_frame(x=1.25,
                    y=c(PFKdown + 7, PFKdown + 12) \%\% 24,
                        xend=1.25,
                        yend=c(PFKup + 7, PFKup + 12) \%\% 24,
                        what=c("PDH", "GS"))
ciopt
split024intervals(ciopt)
}
