% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwutilityfunctions.R
\name{var_value_string}
\alias{var_value_string}
\title{Converts parameters to a string of par__value pairs}
\usage{
var_value_string(...)
}
\arguments{
\item{...}{}
}
\value{
string of par__value pairs separated by ___
}
\description{
Converts parameters to a string of par__value pairs
}
\examples{
a <- 1
b <- 2
var_value_string(a, b)
}
