% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwutilityfunctions.R
\name{ztct_string}
\alias{ztct_string}
\title{Convert Zeitgeber Time Points to Character Strings with Custom Prefix}
\usage{
ztct_string(time_points, prefix = "ZT")
}
\arguments{
\item{time_points}{A numeric vector of Zeitgeber time points.}

\item{prefix}{A character string to prefix each time point. Defaults to "ZT".}
}
\value{
A character vector with formatted Zeitgeber strings.
}
\description{
Translates numeric Zeitgeber time points into character strings formatted as "\{prefix\}\<time\>_\<replicate\>".
}
\examples{
ztct_string(c(0, 0, 2, 2, 4, 4))
# Returns: "ZT0_1" "ZT0_2" "ZT2_1" "ZT2_2" "ZT4_1" "ZT4_2"

ztct_string(c(0, 0, 2, 2, 4, 4), prefix = "CT")
# Returns: "CT0_1" "CT0_2" "CT2_1" "CT2_2" "CT4_1" "CT4_2"

}
